Updated Slave and Client
I added some more code so that in the mean time I can try and connect the two raspberry pis while I dont have hayden's code. The added code brings up a prompt and creates the socket connection to the slave. The slave receives the information.

What needs to be implemented is Hayden's code, Printer connection, and a display.